Refactor: Reorganize directory tree within Rating project
