+++ src/projects/rating/assets/CloseIcon.tsx
