+++ src/projects/rating/assets/StarIcon.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from '../types/IconProps';
+import { IconProps } from '../../../types/IconProps';
 
 export default function StarIcon({ size, fill }: IconProps) {
   return size === 'lg' ? (
