+++ src/projects/rating/assets/ratingImg1.png
