+++ src/projects/rating/components/RatingInput.tsx
