+++ src/projects/rating/lib/constants.ts
