+++ src/projects/rating/lib/type.ts
