+++ src/projects/rating/lib/util.ts
