+++ src/projects/rating/styles/RatingStyle.ts
